django/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
black==23.9.1
cffi==1.15.1
click==8.1.7
colorlog==6.7.0
Django==4.2.5
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
